77958980/requirements.txt
uvicorn==0.27.0.post1
fastapi==0.109.2
python-dotenv==1.0.1
peewee==3.17.1
PyMySQL==1.1.0
cryptography==42.0.2
